versao 1.0

+++ server/src/controllers/professor-controller.ts
@@ -12,13 +12,13 @@ export async function getAllProfessores(req: Request, res: Response, db: mongoos
         response = { data: professores }
     }
     catch (error) {
-        response = { error: error.valueString() }
+        response = { error: "Algo deu errado" }
     }
     return res.send(response)
 }
 
 async function findProfessores(db: mongoose.Db): Promise<ProfessorPublic[]> {
-    const professores: ProfessorPublic[] = await db.collection('professores').find({code:{$exists:false}}).project({ _id: 0, email: 1, nome: 1 }).toArray()
+    const professores: ProfessorPublic[] = await db.collection('professores').find({ code: { $exists: false } }).project({ _id: 0, email: 1, nome: 1 }).toArray()
     return professores;
 }
 
@@ -29,7 +29,7 @@ export async function getProfessorCode(req: Request, res: Response, db: mongoose
         response = { data: professor }
     }
     catch (error) {
-        response = { error: error.valueString() }
+        response = { error: "Algo deu errado" }
     }
     return res.send(response)
 }
@@ -58,7 +58,7 @@ export async function updateProfessorCode(req: Request, res: Response, db: mongo
         response = { data: professor }
     }
     catch (error) {
-        response = { error: error.valueString() }
+        response = { error: "Algo deu errado" }
     }
     return res.send(response)
 }
@@ -70,12 +70,16 @@ async function updateProfessor(email: string, code: string, db: mongoose.Db): Pr
 
 export async function rmvProfessor(req: Request, res: Response, db: mongoose.Db) {
     let response: RequestResult
+    const prof = await db.collection('professores').findOne({ email: req.user.email });
+    if (prof.type != "administrador") {
+        return res.send({ error: "Você não possui permissão para essa operação" });
+    }
     try {
         await deleteProfessor(req.query.email as string, db)
-        response = { data: {succes:"excluído"} }
+        response = { data: { succes: "excluído" } }
     }
     catch (error) {
-        response = { error: error.valueString() }
+        response = { error: "Algo deu errado" }
     }
     return res.send(response)
 }
@@ -94,11 +98,12 @@ export async function login(req: Request, res: Response, db: mongoose.Db) {
     if (!passwordMatched) {
         return res.status(403).send({ error: "Combinação email/senha incorreta" });
     }
+    const SECONDS = 7 * 24 * 60 * 60
     const token = sign({}, auth.jwt.secret, {
         subject: user.email,
         expiresIn: auth.jwt.expiresIn,
     });
-    return res.send({ data: { user, token } })
+    return res.send({ data: { user, token, expiresIn: SECONDS } })
 }
 
 async function findProfessorPrivate(email: string, db: mongoose.Db): Promise<Professor> {
@@ -106,4 +111,18 @@ async function findProfessorPrivate(email: string, db: mongoose.Db): Promise<Pro
     return professor;
 }
 
-
+export async function signup(req: Request, res: Response, db: mongoose.Db) {
+    const userReq = req.body;
+    const user = await findProfessorPrivate(userReq.email as string, db);
+    if (user) {
+        return res.status(401).send({ error: "Uma conta com esse email já existe" });
+    }
+    const invite = await db.collection('invites').findOne({ email: userReq.email });
+    if (!invite || invite.token != userReq.token) {
+        return res.status(403).send({ error: "token de acesso inválido" });
+    }
+    userReq["type"] = "professor";
+    userReq["password"] = await hash(userReq["password"], 12);
+    await db.collection('professores').insertOne(userReq)
+    return res.send({ data: { user } });
+}
